requirements.txt
newsapi-python
requests
flask
boto3
